fix:  rename the NavBar root folder
