+++ src/shared/NavBar/styles.tsx
